requirements.txt
fastapi
uvicorn[standard]
httpx
sse-starlette
pydantic>=2
chromadb
numpy
python-dotenv
orjson
docker
sseclient-py 
requests